requirements.txt
asn1crypto==1.5.1
boto3==1.42.36
botocore==1.42.36
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.4
filelock==3.20.3
idna==3.11
jmespath==1.1.0
kafka-python==2.3.0
numpy==2.4.1
packaging==26.0
pandas==3.0.0
platformdirs==4.5.1
pyarrow==23.0.0
pycparser==3.0
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
snowflake-connector-python==4.2.0
sortedcontainers==2.4.0
tomlkit==0.14.0
typing_extensions==4.15.0
urllib3==2.6.3
